mesa_llm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.0
google-generativeai==0.3.0
mesa==2.2.4
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
networkx==3.1
aiofiles==23.2.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.0